Implement Step 5
A slow but plausibly correct implementation of Step 5.

It tests every combination of discounts and item<->discount relationships and finds the one with the highest savings.

It correctly handles situations where small or medium parcels should be pulled into a mixed-parcel mania.  These would be missed by a simpler (but way faster) algorithm that always forced small and medium parcels into small- and medium-parcel mania discounts.

It’s slow mainly because evaluating the item<->discount relationships is O(N!).  There are potentially smarter ways to do this, like organizing groups based on cost, but I’m currently not confident they are always correct.